requirements.txt
python-telegram-bot==21.4
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
tenacity==9.0.0
web3==7.11.0
eth-account==0.13.7
eth-abi==5.2.0
requests==2.32.3
